+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New command to install app/system on device
 - New build task to build and install in one action
 - Add filters to the package view
+- Add Leaf Profiles view + commands to remove packages from profiles, and to remove profiles
 
 ## 0.3.0
 - All commands reworked to be functional from the command palette